feat(sequencer/feeds_registry): App state added to store feeds registry
Added implementation for registering feed reports and processing
them when the slot time interval is completed.
Added implementation for processing post requests.
When a vote is received as an http POST request, a check is
performed for relevance to the current voting round.
Added trait for feed processing types.
Average processor implemented for feed reports.