feat(docs): Move deployments instructions to root
\# What?

This PR moves the deployment instructions for the node software from the
README of the docs project to the root of the monorepo. It also links
back and forth between the two, to improve discovery for users.

\# Background

This PR is due to the following comment:
https://github.com/blocksense-network/blocksense/pull/338#discussion_r1736911732

tl;dr: Pesho has noted that the right place for the deployment instructions is in the root README.md, not in apps/docs.blocksense.network/README.md.

This was noticed by Toni before, when he suggested the doc fix that
sparked Pesho's comment. We decided to delay the move then, to keep the
fix simpler and easier to review.