feat(data-feeds-config-generator): Add `getOrCreateArtifact` function
If `process.env.ENABLE_CACHE` is true, it memoizes execution of tasks by
serializing their output as JSON files in the artifacts dir on first
execution, and on subsequent ones, loading the respective JSON file (if
it exists) and deserializing it with Effect/Schema. If the memoization
artifact file does not exist, does not contain valid JSON, or fails to
deserialize with the provided schema, the task will be executed again,
as if it wasn't memoized.