feat(sequencer): Upgrade alloy dep to v0.1.1
This is the first tagged version of alloy. It is newer than the revision
that we have locked the dependency at before. We can probably also not
depend on the github repo, but on the actual published crate, if there
is one.

TODO: upgrade to the newest version possible (not trivial)

Tested via `cargo test && cd apps/sequencer && cargo test`.