fix(sequencer): Finally track down stuck transactions
What was the problem?
===

In case a slot contained errors for all feeds from reporters, this lead
to a situation where the errors are never cleared in the map of results.

Solution
===

Copy contents of map when populating list of updates. Clear the map,
before checking whether the list of updates is empty. Only then,
conditionally skip some code, if the list is empty.

Before, the clearing step was erroneously skipped, which lead to the
bug.

Co-authored-by: Hristo Staykov <izo8611@gmail.com>
Co-authored-by: Dimo99 <dimodimov1999@gmail.com>
Co-authored-by: Antoni Dikov <antoni.dikov@gmail.com>