fix(DFCG): Fix problem with `getFieldFromAggregatedData` and add test
While testing `getBaseQuote` noticed that `getFieldFromAggregatedData`
does not work correctly for the `pair` field when the data there is
aggeregated. That was due to the fact that the aggeregated fields are
arrays and we were not handling this case. The aggeregated fields could
be in the following format: `["", ""]`, which is evaluated to true, but
we would like to find field where the pair (the values of in array) are
truty.