fix(sequencer/aggregation_consensus): Fix for awaiting tx futures with futures::select
Futures stored in FuturesUnordered container must be awaited by futures::select instead of directly with tokio::select. On the other hand, the futures that process the rest of the logic in aggregate_batch_consensus_processor are not directly compatible with futures::select. This is a temporary fix that provides a working solution. A better approach is in research.