feat(spin_manifest): Temporarily pass in `additional_feed_info` to oracle scripts
The long term plan is to pass in an unstructured json object that the
oracle script knows how to parse. It is currently named `arguments` in
the `additional_feed_info` section of the feed config, which is
dependent on the `type` field. We will only pass `arguments` when we
move it to the top level of the feed config.