fix(oracles/gecko-terminal): Fix settings parsing
One part of the changes is ensuring that the oracle works with the
updated version of the data in `spin.toml`, which is generated from
`feeds_config_v2`.

However, there is a more complex issue. The check
`if feed_setting.data.contains("pool")` is a temporary solution to
address the following problem:

When running the oracle locally with the command
`just start-oracle gecko-terminal`,
the `apps/oracles/gecko-terminal/spin.toml` file is
read, and the workflow functions as expected. But when running oracle
with our `process-compose` or even in production, we do not have
separate `spin.toml` files for each oracle. Instead, all feed data is
stored in the same file (in local test environments, this is
`.devenv/state/blocksense/reporter/a/spin.toml`). As a result, the
previous approach of deserializing the settings led to errors like
"Can't parse GeckoTerminalPool" because most feeds couldn't process the
`arguments` deserializing due to mismatched schemas.

To solve this, we now check for the string "pool" in the feed data,
which is unique to the `gecko-terminal` oracle. This allows us to safely
decode the arguments.

As mentioned, this is a temporary solution. We should eventually
implement a way to differentiate feed data based on the oracles that
handle them. This was already communicated ( or soon will be ) with
team members.

cc: @PetarKirov.